backend/requirements.txt
# Django Backend for Insight Navigator

# Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0

# Authentication
PyJWT>=2.8.0

# AI Integration
openai>=1.3.0

# File Processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Supabase Integration
supabase>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Production
gunicorn>=21.0.0
whitenoise>=6.6.0

# Development
pytest>=7.4.0
pytest-django>=4.7.0
